+++ task02/ChipboardTablesFacility/Machines/Machines/MachineForChipboard.cs
@@ -6,7 +6,7 @@ namespace Facility.Machines
     public class MachineForChipboard : IMachineForTableLeg, IMachineForTableTop
     {
         private int _priceForSm3 = (int)Material.Chipboard;
-        public TableLeg GetTableLeg(WorkPiece workPiece, int height, int width, int length)
+        public TableLeg GetTableLeg(WorkPiece workPiece, double height, double width, double length)
         {
             if (workPiece.Height < height && workPiece.Width < width && workPiece.Length < length)
                 throw new Exception("Work piece is too small for this detail");
@@ -18,7 +18,7 @@ namespace Facility.Machines
             return new TableLeg(width * length, height, workPiece.Material, _priceForSm3);
         }
 
-        public TableTop GetTableTop(WorkPiece workPiece, int height, int width, int length)
+        public TableTop GetTableTop(WorkPiece workPiece, double height, double width, double length)
         {
             if (workPiece.Height < height && workPiece.Width < width && workPiece.Length < length)
                 throw new Exception("Work piece is too small for this detail");
