+++ task02/ChipboardTablesFacility/Machines/Machines/MachineForMetal.cs
@@ -14,7 +14,7 @@ namespace Facility.Machines
         /// <param name="width"></param>
         /// <param name="length"></param>
         /// <returns>A table leg</returns>
-        public TableLeg GetTableLeg(WorkPiece workPiece, int height, int width, int length)
+        public TableLeg GetTableLeg(WorkPiece workPiece, double height, double width, double length)
         {
             workPiece.Width -= width;
             workPiece.Length -= length;
@@ -33,7 +33,7 @@ namespace Facility.Machines
         /// <param name="height"></param>
         /// <param name="radius"></param>
         /// <returns>A table leg</returns>
-        public TableLeg GetTableLeg(WorkPiece workPiece, int height, int radius)
+        public TableLeg GetTableLeg(WorkPiece workPiece, double height, double radius)
         {
             var diagonal = 2 * radius;
